+++ test/wine_wrapper_test.dart
@@ -8,8 +8,8 @@
 
 import 'dart:io';
 
-import 'package:flutter_win/src/build/wine_wrapper.dart';
-import 'package:flutter_win/src/toolchain.dart';
+import 'package:flutter_build/src/build/wine_wrapper.dart';
+import 'package:flutter_build/src/toolchain.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';